requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
crispy-bootstrap4==2022.1
Django==4.1.1
django-bootstrap4==23.1
django-countable-field==1.3
django-crispy-forms==2.0
Pillow==9.5.0
soupsieve==2.4.1
sqlparse==0.4.2
